requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
gspread==6.1.0
google-auth==2.36.0
jinja2==3.1.4
python-multipart==0.0.18
httpx==0.28.0
pydantic-settings==2.6.0
